fixed resolution of documents with deep references
Note to self:

When encountering a situation where it isn't clear how to get the data
from sanity, check to see if the document you are referencing is too
deep in the tree for the resolvers.

In this case I just doubled the maxDepth to 10, but it's possible this
may not suffice for everything.