Finished Refactoring Contact form to use Formik!
This was a major update and improvement over the old very janky way I
was doing things.

The main components are using Formik, Form and useField to prevent us
from having to really deal with any of the state changes.

validation is done through Yup. I'm keeping the StyledErrorMessage for
now (which currently is set to {display: none;}) in the event that our
client wants to add something more than the coloring and shake effect.

Overall, while the code isn't particularly shorter than the old way -
it's way easier to reason through!

TODO: hook up navigation to /thank-you and y'know actually submit the
form.